+++ packages/framework/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @cosmicdrift/kumiko-framework
 
+## 0.37.0
+
 ## 0.36.0
 
 ### Minor Changes