+++ packages/headless/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @cosmicdrift/kumiko-headless
 
+## 0.37.0
+
+### Patch Changes
+
+- @cosmicdrift/kumiko-framework@0.37.0
+
 ## 0.36.0
 
 ### Minor Changes