+++ packages/renderer-web/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @cosmicdrift/kumiko-renderer-web
 
+## 0.37.0
+
+### Patch Changes
+
+- @cosmicdrift/kumiko-dispatcher-live@0.37.0
+- @cosmicdrift/kumiko-headless@0.37.0
+- @cosmicdrift/kumiko-renderer@0.37.0
+
 ## 0.36.0
 
 ### Patch Changes