+++ packages/renderer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @cosmicdrift/kumiko-renderer
 
+## 0.37.0
+
+### Patch Changes
+
+- @cosmicdrift/kumiko-framework@0.37.0
+- @cosmicdrift/kumiko-headless@0.37.0
+
 ## 0.36.0
 
 ### Patch Changes